+++ src/api/category/category.controller.ts
@@ -2,7 +2,7 @@ import { Controller, Get, UseGuards } from '@nestjs/common';
 
 import { JwtAuthGuard } from '@src/shared/guard/jwt-auth.guard';
 
-import { CategoryService } from '../service/category.service';
+import { CategoryService } from './service/category.service';
 import { CategoryShowResponse } from './dto/response/category-show.response';
 
 @UseGuards(JwtAuthGuard)
